Refactor Sage Terminal modules
